Создал отдельный виджет для кнопки добавления задачи
