+++ lib/features/home/widgets/widgets.dart
@@ -2,6 +2,7 @@
 // generated by: https://marketplace.visualstudio.com/items?itemName=elmsec.dart-flutter-exports:
 
 
+export 'add_task_button.dart';
 export 'add_task_dialog.dart';
 export 'delete_task_dialog.dart';
 export 'tasks_list.dart';
